Use CRTP for shape polymorphism
Sphere does not contain an object anymore, but is an object. This is
done with CRTP as vtables created during instantation on the host will
contain pointers to host memory. The other alternative is to create the
polymorphic objects in a kernel. That was deemed to be a much larger job
though. Lets hope this will work for future features as well :).