CRTP to the maxx
CRTP is now used for materials and objects. Some circular dependencies
though due to materials needing access the the Tracer to cast new rays.
Could extract an interface that the Tracer implements to break the
circular dependency but I'm probably leaving this trail here. To much
effort compared to other alternatives.