Support bumpmapping/normalmapping
Some visual artifacts in the computed normal though. T, B and N looks
okay when rendered directly. When the composite bumpmapped surface is
computed though the normals looks strange at some vertices. When
compared to a rasterised DirectX application with the same algorithm,
this does not happen. Will ahve to leave as is as I'm not getting any
further right now.