Some refactoring
Split tracer.cu and cuh into oen for the class Tracer and one for the
entry functions to cuda (now moved to main.cu/cuh).

Sometimes compilation fails due to there being two .cu-files. Shoudl be
fixed by adding the /FS flag to the host-side compilation flags given to
nvcc. Thats for a future commit though.

Ray moved to own file in preparation for being used in the cuda code,
afirst step on the BVH journey :)