Use BVH in raytracer, albeit with a visual bug
Not all triangles that should be rendered are rendered. The current
theory is that triangle bouynding boxes are not correctly calculated or
that triangles end up in the wrong node. Need to add better
visualisation of bounding boxes or more unit tests to debug properly.

Some commented away code still remains as debugging is in progress.