Solve visual bug
The boundingbox constructor taking a triangle and vertices was faulty.
bbmin and bbmax was initialized to 0 for the Y and Z axes. Only the X
axius was properly initialized to FLT_MIN and FLT_MAX. This caused
centroids to be off which in turn meant that triangles ended up outside
of their bounding boxes.

Theres a lot of commented out and non out-commented debug code left :).
Will clean that up in following commits and also add back the plane.