Replace hard coded transformation with matrix
A matrix can be supplied to a mesh. That matrix will be inversed and the
inverse will be used to transform the ray for intersection tests. The
vertices themselves does not need to be manipulated in any way and model
reuse is easier .This is of course not free performance wise but as the
inverse is precalculated it not too expensive.