Add more Options and support separate mesh render
More options are now visible in teh debug GUI. Some options have been
moved to their own structs.

An ungly hack to debug has been added, If a model is specified in the
options, only that will be rendered. Ideally, the scene should be
constructed outside the RayTracer class and submitted through the
constructor.