Fix PicDump bug and change experiment setup
PicDump didn't work as the array it was allocating was of the wrong size

The experiments are now loaded from json files and are not defined in
code. The reason for this is to make sure that each expirement is
executed from a clean slate. There should be no risk of uncleaned, old
memory, laying around on the device or host. This cahnge is to mitigate
this risk.